Code refactoring. Structure modification
